[PDB] Fix a bug where we were serializing hash tables incorrectly.
There was some code that tried to calculate the number of 4-byte
words required to hold N bits, but it was instead computing the
number of bytes required to hold N bits.  This was leading to
extraneous data being output into the hash table, which would
cause certain operations in DIA (the Microsoft PDB reader) to
fail.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@327675 91177308-0d34-0410-b5e6-96231b3b80d8