[ShrinkWrap] Take into account landing pad
When scanning the function for CSRs uses and defs, also check if
the basic block are landing pads.
Consider that landing pads needs the CSRs to be properly set.
That way we force the prologue/epilogue to always be pushed out
of the problematic "throw" region. The "throw" region is
problematic because the jumps are not properly modeled.

Fixes PR36513

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@327942 91177308-0d34-0410-b5e6-96231b3b80d8