.debug_names: Parse DW_IDX_die_offset as a reference
Before this patch we were parsing the attributes as section offsets, as
that is what apple_names is doing. However, this is not correct as DWARF
v5 specifies that this attribute should use the Reference form class.

This also updates all the testcases (except the ones that deliberately
pass a different form) to use the correct form class.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@328773 91177308-0d34-0410-b5e6-96231b3b80d8