[AArch64] Fix regression after r329691
In r329691, we would choose FP even if the offset wouldn't fit, just
because the offset is smaller than the one from BP. This made many
accesses through FP need to scavenge a register, which resulted in
slower and bigger code for no good reason.

This patch now always picks the offset that fits first, even if FP is
preferred.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@329797 91177308-0d34-0410-b5e6-96231b3b80d8