[RuntimeDyld] Add an assert to catch misbehaving symbol resolvers.
Resolvers are required to find results for all requested symbols or return an
error, but if a resolver fails to adhere to this contract (by returning results
for only a subset of the requested symbols) then this code will infinite loop.
This assertion catches resolvers that fail to adhere to the contract.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@334536 91177308-0d34-0410-b5e6-96231b3b80d8