[docs] Switch debug info metadata blocks to use `text` instead of `llvm` highlighting syntax.
Most of them already were like this, and the Sphinx runs on the docs
build bot seems to be substantially more picky and/or not have support
for a bunch of the syntax here. Hopefully this will let it progress past
this.

My previous attempt to fix the syntax made the `opt` tool happy, but no
idea what the Sphinx stuff is really looking for, and the fact that
other blocks already just use `text` led me to this solution.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@338983 91177308-0d34-0410-b5e6-96231b3b80d8